feat: move depreacted `Loader` to `compat` package & replace with `BusyIndicator`
