+++ packages/charts/src/internal/ChartContainer.tsx
@@ -1,7 +1,7 @@
-import { type CommonProps, Label, Loader } from '@ui5/webcomponents-react';
+import { type CommonProps, Label } from '@ui5/webcomponents-react';
 import { useStylesheet } from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
-import type { ComponentType, CSSProperties, ReactElement, ReactNode } from 'react';
+import type { ComponentType, ReactElement, ReactNode } from 'react';
 import { Component, forwardRef } from 'react';
 import { ResponsiveContainer } from 'recharts';
 import { classNames, styleData } from './ChartContainer.module.css.js';
@@ -13,13 +13,6 @@ export interface ContainerProps extends CommonProps {
   loading?: boolean;
   resizeDebounce: number;
 }
-
-const loaderStyles: CSSProperties = {
-  position: 'absolute',
-  top: 0,
-  left: 0,
-  right: 0
-};
 
 class ErrorBoundary extends Component<{ children: ReactNode }, { errorCount: number }> {
   state = {
@@ -49,7 +42,8 @@ const ChartContainer = forwardRef<HTMLDivElement, ContainerProps>((props, ref) =
     <div ref={ref} className={clsx(classNames.container, className)} slot={slot} {...rest}>
       {dataset?.length > 0 ? (
         <>
-          {loading && <Loader style={loaderStyles} />}
+          {/*todo replace with BusyIndicator*/}
+          {loading && 'Loading...'}
           <ErrorBoundary>
             <ResponsiveContainer debounce={resizeDebounce}>{children}</ResponsiveContainer>
           </ErrorBoundary>
