+++ packages/compat/src/enums/LoaderType.ts
