+++ packages/compat/src/index.ts
@@ -3,3 +3,6 @@ export * from './components/TableCell/index.js';
 export * from './components/TableColumn/index.js';
 export * from './components/TableGroupRow/index.js';
 export * from './components/TableRow/index.js';
+export * from './components/Loader/index.js';
+
+export { LoaderType } from './enums/LoaderType.js';
