--- packages/main/src/components/AnalyticalTable/defaults/LoadingComponent/DefaultLoadingComponent.module.css
@@ -1,7 +1,0 @@
-.loader {
-  position: absolute;
-  inset-block-start: --_ui5wcr-AnalyticalTableHeaderRowHeight;
-  inset-inline-start: 0;
-  inset-inline-end: 0;
-  z-index: 1;
-}
