--- packages/main/src/components/Loader/Loader.module.css
@@ -1,41 +1,0 @@
-.loader {
-  position: relative;
-  height: 0.25rem;
-  width: 100%;
-
-  &:before {
-    content: '';
-    position: absolute;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: var(--sapContent_BusyColor);
-    opacity: 0.15;
-  }
-
-  &.loaderDeterminate {
-    background: linear-gradient(to right, var(--sapContent_BusyColor), var(--sapContent_BusyColor)) repeat-y;
-  }
-
-  &.loaderIndeterminate {
-    background-size: 40%;
-    background: linear-gradient(
-        to right,
-        transparent 0px,
-        var(--sapContent_BusyColor) calc(50% - 2rem),
-        var(--sapContent_BusyColor) calc(50% + 2rem),
-        transparent 100%
-      )
-      repeat-y;
-    animation: scroll 1.2s linear infinite;
-  }
-}
-
-@keyframes scroll {
-  0% {
-    background-position: -100% 0;
-  }
-  100% {
-    background-position: 200% 0;
-  }
-}
